requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
peewee==3.13.1
python-dateutil==2.8.1
PyYAML==5.3
pyzmq==19.0.0
six==1.14.0
websockets==8.1
yarl==1.4.2
